bratulek/requirements.txt
Django==1.8
django-ckeditor==5.1.1
Pillow==3.3.1
